requirements/dev.txt
aiohttp==3.8.5
aioredis==1.3.1
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.103.1
fastapi-cache2==0.2.1
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
hiredis==2.2.3
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
orjson==3.9.7
passlib==1.7.4
pendulum==2.1.2
pycparser==2.21
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytzdata==2020.1
PyYAML==6.0.1
redis==5.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
